+++ generators/python/src/fern_python/generators/pydantic_model/type_declaration_handler/discriminated_union/simple_discriminated_union_generator.py
@@ -355,7 +355,7 @@ class DiscriminatedUnionSnippetGenerator:
         if isinstance(example, ir_types.ExampleObjectTypeWithTypeId):
             args.extend(
                 self.snippet_writer.get_snippet_for_object_properties(
-                    example=example.object,
+                    example=example.object, request_parameter_names={}
                 ),
             )
         else: