+++ generators/python/src/fern_python/generators/pydantic_model/type_declaration_handler/object_generator.py
@@ -79,7 +79,7 @@ class ObjectSnippetGenerator:
                     name=self.name,
                 ),
                 args=self.snippet_writer.get_snippet_for_object_properties(
-                    example=self.example,
+                    example=self.example, request_parameter_names={}
                 ),
             ),
         )