+++ generators/python/src/fern_python/generators/sdk/client_generator/request_body_parameters/abstract_request_body_parameters.py
@@ -1,5 +1,7 @@
 from abc import abstractmethod
-from typing import List, Optional
+from typing import Dict, List, Optional
+
+import fern.ir.resources as ir_types
 
 from fern_python.codegen import AST
 
@@ -27,4 +29,8 @@ class AbstractRequestBodyParameters:
 
     @abstractmethod
     def get_content(self) -> Optional[AST.Expression]:
+        ...
+
+    @abstractmethod
+    def get_parameter_name_rewrites(self) -> Dict[ir_types.Name, str]:
         ...