+++ seed/csharp-sdk/multi-url-environment/src/SeedMultiUrlEnvironment/S3/requests/GetPresignedUrlRequest.cs
@@ -1,6 +1,9 @@
+using System.Text.Json.Serialization;
+
 namespace SeedMultiUrlEnvironment;
 
 public class GetPresignedUrlRequest
 {
+    [JsonPropertyName("s3Key")]
     public string S3Key { get; init; }
 }