+++ seed/csharp-sdk/streaming/src/SeedStreaming/Dummy/requests/GenerateStreamRequestzs.cs
@@ -1,6 +1,9 @@
+using System.Text.Json.Serialization;
+
 namespace SeedStreaming;
 
 public class GenerateStreamRequestzs
 {
+    [JsonPropertyName("num_events")]
     public int NumEvents { get; init; }
 }