+++ seed/csharp-sdk/trace/src/SeedTrace/Admin/requests/StoreTracedTestCaseRequest.cs
@@ -1,10 +1,13 @@
+using System.Text.Json.Serialization;
 using SeedTrace;
 
 namespace SeedTrace;
 
 public class StoreTracedTestCaseRequest
 {
+    [JsonPropertyName("result")]
     public TestCaseResultWithStdout Result { get; init; }
 
+    [JsonPropertyName("traceResponses")]
     public List<TraceResponse> TraceResponses { get; init; }
 }