+++ seed/python-sdk/exhaustive/skip-pydantic-validation/src/seed/endpoints/http_methods/client.py
@@ -390,6 +390,9 @@ class HttpMethodsClient:
                     request_options.get("additional_query_parameters") if request_options is not None else None
                 )
             ),
+            json=jsonable_encoder(remove_none_from_dict(request_options.get("additional_body_parameters", {})))
+            if request_options is not None
+            else None,
             headers=jsonable_encoder(
                 remove_none_from_dict(
                     {
@@ -784,6 +787,9 @@ class AsyncHttpMethodsClient:
                     request_options.get("additional_query_parameters") if request_options is not None else None
                 )
             ),
+            json=jsonable_encoder(remove_none_from_dict(request_options.get("additional_body_parameters", {})))
+            if request_options is not None
+            else None,
             headers=jsonable_encoder(
                 remove_none_from_dict(
                     {