+++ seed/python-sdk/oauth-client-credentials-environment-variables/src/seed/client.py
@@ -13,7 +13,7 @@ from .core.oauth_token_provider import OAuthTokenProvider
 
 class SeedOauthClientCredentialsEnvironmentVariables:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------
@@ -83,7 +83,7 @@ class SeedOauthClientCredentialsEnvironmentVariables:
 
 class AsyncSeedOauthClientCredentialsEnvironmentVariables:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------