+++ seed/python-sdk/package-yml/src/seed/client.py
@@ -21,7 +21,7 @@ OMIT = typing.cast(typing.Any, ...)
 
 class SeedPackageYml:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------
@@ -127,7 +127,7 @@ class SeedPackageYml:
 
 class AsyncSeedPackageYml:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------