+++ seed/python-sdk/unions/src/seed/client.py
@@ -10,7 +10,7 @@ from .union.client import AsyncUnionClient, UnionClient
 
 class SeedUnions:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------
@@ -58,7 +58,7 @@ class SeedUnions:
 
 class AsyncSeedUnions:
     """
-    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propagate to these functions.
 
     Parameters
     ----------